refactor: separar footer em um aplicativo
