+++ apps/footer/apps.py
@@ -1,5 +1,6 @@
 from django.apps import AppConfig
 
-class AppsConfig(AppConfig):
+
+class FooterConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps'
+    name = 'apps.footer'
