+++ apps/footer/urls.py
@@ -1,0 +1,2 @@
+urlpatterns = [
+]
