+++ manicursos/urls.py
@@ -22,7 +22,7 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('apps.urls')),
+    path('', include('apps.footer.urls')),
 ]
